r2337/sofa-dev : Fix : trying to fix warning for gcc 4.0/4.1 and maybe the errors for mac os x

+++ modules/sofa/component/visualmodel/OglVariable.cpp
@@ -289,7 +289,7 @@ int OglFloatVector4VariableClass = core::RegisterObject("OglFloatVector4Variable
         ;
 
 OglFloatVectorVariable::OglFloatVectorVariable()
-    : fv(initData(&fv, (helper::vector<float>) helper::vector<float>(1.0), "values", "Set float array values"))
+    : fv(initData(&fv, (helper::vector<float>) helper::vector<float>((float)1.0), "values", "Set float array values"))
 {
 
 }
